more local storage persistence + made changes to nav based on samuel and jamie feeback
