+++ package.json
@@ -14,6 +14,9 @@
     "dependencies": {
         "@emotion/react": "^11.11.1",
         "@emotion/styled": "^11.11.0",
+        "@fortawesome/fontawesome-svg-core": "^6.5.1",
+        "@fortawesome/free-solid-svg-icons": "^6.5.1",
+        "@fortawesome/react-fontawesome": "^0.2.0",
         "@imc-trading/react-pivottable": "^0.2.13",
         "@mui/icons-material": "^5.14.19",
         "@mui/material": "^5.14.20",
