+++ src/routes/Navbar.css
@@ -15,9 +15,7 @@
     font-weight: 500;
 }
 
-.nav-link {
-    font-size: 16px;
-}
+.nav-link {}
 
 .navbar-brand {
     margin-left: 1rem;
