+++ src/routes/Sidebar.css
@@ -1,25 +1,47 @@
 .sidebar {
     height: 100%;
-    width: 150px;
-    /* Width of the sidebar */
+    width: 200px;
+    /* Adjust width as needed */
     position: fixed;
     z-index: 1;
     top: 0;
     left: 0;
-    background-color: #111;
+    background: linear-gradient(90deg, rgb(28, 27, 27) 0%, rgb(26, 23, 23) 100%);
     overflow-x: hidden;
     padding-top: 20px;
-    margin-right: 10px;
+    padding-left: 10px;
+    /* Add some padding to the left */
 }
 
 .sidebar a {
-    padding: 6px 8px 6px 16px;
+    padding: 10px;
+    /* Adjust padding for the size of the rectangle */
     text-decoration: none;
-    font-size: 25px;
-    color: #818181;
-    display: block;
+    color: white;
+    font-family: 'Roboto', sans-serif;
+    font-weight: 500;
+    display: flex;
+    align-items: center;
+    font-size: 20px;
+    margin-right: 10px;
+    /* Adds margin around the links for space from the edges */
+    border-radius: 10px;
+    /* Rounded edges for the rectangle */
+    transition: background-color 0.3s;
+    /* Smooth transition for background color */
 }
 
 .sidebar a:hover {
-    color: #f1f1f1;
+    color: white;
+    /* Keep text color white on hover */
+    background-color: #757575;
+    /* Grey background on hover */
+}
+
+.sidebar a svg {
+    margin-right: 10px;
+    margin-left: 5px;
+    width: 20px;
+    /* Adjust size as needed */
+    height: auto;
 }
