--- src/views/Feedback.css
@@ -1,26 +1,0 @@
-.feedback-textarea {
-    height: 100px;
-    /* Adjust the height as needed */
-}
-
-.chat-container {
-    margin-top: 0;
-    width: 85vw;
-    height: 100%;
-    border-radius: 4px;
-    margin: 1rem;
-    margin-left: 90px;
-    padding: 2rem;
-    justify-content: center;
-    /* or whatever value you desire for the chatbot page */
-}
-
-.feedback-container {
-    margin-top: 0;
-    height: 100%;
-    border-radius: 4px;
-    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
-    padding: 2rem;
-    border: 1px dashed #000;
-    justify-content: center;
-}
