+++ src/views/Library.css
@@ -1,5 +1,4 @@
 .library-container {
-    display: flex;
     /* Align the tops of the two children */
 }
 
@@ -12,7 +11,6 @@
 }
 
 .upload-component {
-    margin-left: 20px;
     /* Spacing between table and upload component */
     display: flex;
     min-height: 50vh;
